cleanString and code in Utils instead

+++ src/commands/devs/eval.js
@@ -1,6 +1,7 @@
 const { Command, CommandError, SimplicityEmbed, Utils } = require('../../')
 const { inspect } = require('util')
-const value = (lang, code) => `\`\`\`${lang}\n${String(code).slice(0, 2045) + (code.length >= 2045 ? '...' : '')}\n\`\`\``.replace(process.env.BOT_TOKEN, () => '*'.repeat(process.env.BOT_TOKEN.length))
+const { code } = Utils
+const value = (lang, str) => code(str, lang).replace(process.env.BOT_TOKEN, () => '*'.repeat(process.env.BOT_TOKEN.length))
 
 class Eval extends Command {
   constructor (client) {
@@ -15,10 +16,10 @@ class Eval extends Command {
   async run ({ author, botLanguages, client, guild, channel, member, language, command, prefix, message, query, send, args, t, emoji }) {
     const embed = new SimplicityEmbed({ author })
 
-    const code = query.replace(/^```(js|javascript ?\n)?|```$/g, '')
+    const text = query.replace(/^```(js|javascript ?\n)?|```$/g, '')
 
     try {
-      const evald = eval(code)
+      const evald = eval(text)
       const toEval = inspect(evald, { depth: 0 })
 
       embed
@@ -45,8 +46,8 @@ class Eval extends Command {
         const collector = await msg.createReactionCollector(filter, { max: 1, errors: ['time'], time: 30000 })
 
         collector.on('collect', async () => {
-          await msg.delete()
-          await message.delete()
+          if (msg) await msg.delete()
+          if (message) await message.delete()
         })
         collector.on('end', async () => {
           if (msg) await msg.reactions.removeAll().catch(() => null)
